Initial Development Begun
I have started the development of the Northwind Traders Complete Accounting Basic Edition. The following modules have been created, with the included classes:

* com_pekinsoft_northwind
    * com.pekinsoft.northwind
        * Startup.java
* com_pekinsoft_northwind_basic
    * com.pekinsoft.northwind.basic
        * Application.java
* com_pekinsoft_northwind_utils
    * com.pekinsoft.northwind.utils
        * ArgumentParser.java
        * FileUtils.java
        * Logger.java
        * MessageBox.java
        * PasswordUtils.java
        * StringUtils.java
        * Utils.java
    * com.pekinsoft.northwind.utils.enums
        * SysExits.java
    * com.pekinsoft.northwind.utils.exceptions
        * InvalidLoggingLevelException.java
        * InvalidPropertiesFormatException.java

The next modules to be created are:

* com_pekinsoft_northwind_accounting - This module will contain all accounting functionality, as well as the General Ledger for all editions of Northwind Traders.
* com_pekinsoft_northwind_basic_desktop - This module will contain all GUI-related functionality for the Basic Edition of Northwind Traders.
* com_pekinsoft_northwind_loads - This module will contain all load-related tracking functionality for all editions of Northwind Traders.
* com_pekinsoft_northwind_vehicles - This module will contain all vehicle/unit and service management functionality for all editions of Northwind Traders.
* com_pekinsoft_northwind_reporting - This module will contain all reporting functionality for all editions of Northwind Traders.

There will be additional packages added to the com_pekinsoft_northwind_basic module that will provide for user and customer management features for all editions of Northwind Traders. I want to try and keep those common packages in a single module for ease of upkeep after release of the application.

After I get the Basic Edition up to a functional level, I'll start working on the Corporate Edition, then the Enterprise Edition. However, I want to be sure all of the basic functionality is in prime working condition, as it will be reused in the other two editions of the application.