Desktop Module Added
I have added the `com_pekinsoft_northwind_desktop` module to the project, which contains the packages:

* `com.pekinsoft.northwind.desktop`: where all window/GUI classes will go.
* `com.pekinsoft.northwind.desktop.graphics`: where all graphics for the system will be placed, in the appropriate subpackage:
    - `com.pekinsoft.northwind.desktop.graphics.x16`: 16 x 16 icons
    - `com.pekinsoft.northwind.desktop.graphics.x32`: 32 x 32 icons
    - `com.pekinsoft.northwind.desktop.graphics.x48`: 48 x 48 icons
    - etc. Only `x16` subpackage has been created and only two icons have been added:
        * `exit.png`
        * `Northwind.png

I have also created two issues (#3 and #4) dealing with what I have worked on so far. Issue #3 is for feedback on the log entries, and Issue #4 is for assistance in getting automatic version incrementation working.