Autoincrementing Versioning System Working
I figured out what was wrong with the autoincrementing of the Build number. Though I don't exactly know why, the following code did *not* work:

```java
    if ( Boolean.getBoolean(props.getProperty("debugging")) ) {
		// Code to execute during debugging here...
	}
```

However, when I changed it to the following, the build number started incrementing as expected:

```java
	DEBUGGING = Boolean.parseBoolean(props.getProperty("debugging", "false"));

    if ( DEBUGGING == true ) {
		// Code to execute during debugging here...
	}
```

Now, as long as we are running the application from NetBeans in development/debugging mode, we will have an incrementing Build number. Once the build reaches 9999, the next time we debug it will change to zero (0) and the Revision will increment by one (1).

I am going to close Issue #4 with this commit.