Started Creating Dialogs
I have started creating dialogs, though I have only created one so far, `CustomerEntryDialog`. I created it as an `InternalJFrame` form so that we can keep all windows/dialogs inside of the `MainFrame` form. Since I am a visual person, I tend to think visually, so I will create the dialogs in order to give me the perspective I need on creating the tables in the database. I will also create the flow, as in how each dialog relates to the others, which is how the tables will relate to each other in the database.