LoadEntryDialog Created
I have created the `LoadEntryDialog` as a `JInternalFrame` class to receive all information about the load being dispatched. This dialog is created in the `com.pekinsoft.northwind.desktop` package and provides a button to add a stop to the load, with a table to store all stops for the load. With the creation of the `LoadEntryDialog`, I have added a new `JMenuItem` under the `File...New...` `JMenu` structure for entering the load information. Furthermore, I have assigned the keyboard accelerator of `F5` to pull up the `LoadEntryDialog` as well.

### Database Notes:
The LOADS table will have a one-to-many relationship with the STOPS table. The STOPS table will have a many-to-one relationship with the CUSTOMERS table. These relationships will need a ***cascade*** referential integrity set up on them to make sure that no record may be deleted in the 'one' side of the relationship has any relatives in the 'many' side of the relationship.

### StopEntryDialog Notes:
When I create the `StopEntryDialog`, there will be a table listing all customers in the database so that the user may select the customer for the stop. There will also be a "filter" option and a "search" option to narrow the list as the CUSTOMERS table grows with data. Furthermore, I will provide a way for a new customer entry to be added to the CUSTOMERS table from the `StopEntryDialog` so that the user will be able to also add customers when they are needed, after searching for the customer in the CUSTOMERS table and find that the customer has not yet been added to the database.