StopEntryDialog Created
I have created the `StopEntryDialog` for allowing the user to select an existing customer as a stop off for the load. There is a button in the bottom-left corner that allows the user to bring up the `CustomerEntryDialog` if the customer they are looking for does not yet exist in the database. I have also provided for filtering the customers list by `CUST_ID`, Company, City, or State, thereby allowing the user to shorten the list and quickly see if the customer they are looking for exists.