Updated Logger.critical
As explained in my last comment on Issue #9, I have updated the `Logger.critical` method to include the list of installed modules in the error log.

I also added a `manifest.xml` file to the default package of all modules except for `com_pekinsoft_northwind`, as it only contains the `Startup` class.

Also, @jkovalsy, there is a logical issue with the following code:

```java
        // Let's customize the main menu bar by loaded modules
        while (modules.hasNext()) {
            Module module = modules.next();
            MenuProvider menuProvider = (MenuProvider) module.getInstanceOf("com.pekinsoft.northwind.basic.MenuProvider");
            jMenuBar1.add(menuProvider.getMenu(this.MainDesktop));
        }
```

The problem is that, with the other modules in the `${HOME}/.northwind/modules` folder, `NullPointerException`s are thrown if the module does not contain a `menuProvider` object. Since that is your block of code, I thought that I would let you know this and, perhaps, you can add a check for the `menuProvider` within this block.

I am calling it a night. Tomorrow, I will move the `Module` and `ModulesManager` classes back to the `com_pekinsoft_northwind_basic` module, and remove the `com_pekinsoft_northwind_basic_modules` module.